+++ common/src/main/java/hmd/HmdCommonApplication.java
@@ -1,13 +1,13 @@
-package gaia3d;
+package hmd;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DockyardAdminApplication {
+public class HmdCommonApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DockyardAdminApplication.class, args);
+		SpringApplication.run(HmdCommonApplication.class, args);
 	}
 
 }
