+++ dockyard-user/src/main/java/hmd/DockyardUserApplication.java
@@ -1,13 +1,13 @@
-package gaia3d;
+package hmd;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class FacilityUserApplication {
+public class DockyardUserApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(FacilityUserApplication.class, args);
+		SpringApplication.run(DockyardUserApplication.class, args);
 	}
 
 }
