+++ facility-admin/src/main/java/hmd/FacilityAdminApplication.java
@@ -1,13 +1,13 @@
-package gaia3d;
+package hmd;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class GeoUserApplication {
+public class FacilityAdminApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(GeoUserApplication.class, args);
+		SpringApplication.run(FacilityAdminApplication.class, args);
 	}
 
 }
