--- facility-user/src/main/java/gaia3d/ServletInitializer.java
@@ -1,14 +1,0 @@
-package gaia3d;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(FacilityUserApplication.class);
-	}
-
-}
-
