+++ gis-admin/src/main/java/hmd/GisAdminApplication.java
@@ -1,14 +1,14 @@
-package gaia3d;
+package hmd;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
 @SpringBootApplication
-public class GeoAdminApplication extends SpringBootServletInitializer {
+public class GisAdminApplication extends SpringBootServletInitializer {
 
 	public static void main(String[] args) {
-		SpringApplication.run(GeoAdminApplication.class, args);
+		SpringApplication.run(GisAdminApplication.class, args);
 	}
 
 }
