+++ gis-admin/src/test/java/gaia3d/GisAdminApplicationTests.java
@@ -7,7 +7,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class GeoAdminApplicationTests {
+public class GisAdminApplicationTests {
 
 	@Test
 	public void contextLoads() {
